put error and success in right files
