+++ documentdb_tests/compatibility/tests/system/administration/commands/setFeatureCompatibilityVersion/test_setFeatureCompatibilityVersion_writeConcern_accepted.py
@@ -1,7 +1,7 @@
-"""Tests for setFeatureCompatibilityVersion writeConcern field validation.
+"""Tests for setFeatureCompatibilityVersion writeConcern field acceptance.
 
-Validates writeConcern type validation, null-as-omitted behavior,
-empty-doc acceptance, and wtimeout coercion.
+Validates that writeConcern accepts valid object values, null-as-omitted,
+empty-doc, omission, and various numeric types for wtimeout.
 """
 
 import pytest
@@ -12,7 +12,6 @@ from documentdb_tests.compatibility.tests.core.utils.command_test_case import (
     CommandTestCase,
 )
 from documentdb_tests.framework.assertions import assertResult
-from documentdb_tests.framework.error_codes import TYPE_MISMATCH_ERROR
 from documentdb_tests.framework.executor import execute_admin_command
 from documentdb_tests.framework.parametrize import pytest_params
 
@@ -62,82 +61,6 @@ WRITE_CONCERN_ACCEPTED_TESTS: list[CommandTestCase] = [
         },
         expected={"ok": 1.0},
         msg="setFeatureCompatibilityVersion should succeed when writeConcern is omitted",
-    ),
-]
-
-
-# Property [writeConcern Rejected]: setFeatureCompatibilityVersion rejects
-# non-object writeConcern types.
-WRITE_CONCERN_REJECTED_TESTS: list[CommandTestCase] = [
-    CommandTestCase(
-        "string",
-        command=lambda ctx: {
-            "setFeatureCompatibilityVersion": "CURRENT_FCV",
-            "confirm": True,
-            "writeConcern": "majority",
-        },
-        error_code=TYPE_MISMATCH_ERROR,
-        msg="setFeatureCompatibilityVersion should reject writeConcern as string",
-    ),
-    CommandTestCase(
-        "int",
-        command=lambda ctx: {
-            "setFeatureCompatibilityVersion": "CURRENT_FCV",
-            "confirm": True,
-            "writeConcern": 1,
-        },
-        error_code=TYPE_MISMATCH_ERROR,
-        msg="setFeatureCompatibilityVersion should reject writeConcern as int",
-    ),
-    CommandTestCase(
-        "bool",
-        command=lambda ctx: {
-            "setFeatureCompatibilityVersion": "CURRENT_FCV",
-            "confirm": True,
-            "writeConcern": True,
-        },
-        error_code=TYPE_MISMATCH_ERROR,
-        msg="setFeatureCompatibilityVersion should reject writeConcern as bool",
-    ),
-    CommandTestCase(
-        "array",
-        command=lambda ctx: {
-            "setFeatureCompatibilityVersion": "CURRENT_FCV",
-            "confirm": True,
-            "writeConcern": [{"w": 1}],
-        },
-        error_code=TYPE_MISMATCH_ERROR,
-        msg="setFeatureCompatibilityVersion should reject writeConcern as array",
-    ),
-    CommandTestCase(
-        "long",
-        command=lambda ctx: {
-            "setFeatureCompatibilityVersion": "CURRENT_FCV",
-            "confirm": True,
-            "writeConcern": Int64(1),
-        },
-        error_code=TYPE_MISMATCH_ERROR,
-        msg="setFeatureCompatibilityVersion should reject writeConcern as long",
-    ),
-    CommandTestCase(
-        "double",
-        command=lambda ctx: {
-            "setFeatureCompatibilityVersion": "CURRENT_FCV",
-            "confirm": True,
-            "writeConcern": 1.0,
-        },
-        error_code=TYPE_MISMATCH_ERROR,
-        msg="setFeatureCompatibilityVersion should reject writeConcern as double",
-    ),
-    CommandTestCase(
-        "decimal128",
-        command=lambda ctx: {
-            "setFeatureCompatibilityVersion": "CURRENT_FCV",
-            "confirm": True,
-            "writeConcern": Decimal128("1"),
-        },
-        error_code=TYPE_MISMATCH_ERROR,
-        msg="setFeatureCompatibilityVersion should reject writeConcern as Decimal128",
     ),
 ]
 
@@ -201,23 +124,6 @@ WTIMEOUT_COERCION_TESTS: list[CommandTestCase] = [
 @pytest.mark.parametrize("test", pytest_params(WRITE_CONCERN_ACCEPTED_TESTS))
 def test_setFeatureCompatibilityVersion_writeConcern_accepted(database_client, collection, test):
     """Test setFeatureCompatibilityVersion accepts valid writeConcern values."""
-    collection = test.prepare(database_client, collection)
-    ctx = CommandContext.from_collection(collection)
-    cmd = test.build_command(ctx)
-    cmd["setFeatureCompatibilityVersion"] = get_fcv(collection)
-    result = execute_admin_command(collection, cmd)
-    assertResult(
-        result,
-        expected=test.build_expected(ctx),
-        error_code=test.error_code,
-        msg=test.msg,
-        raw_res=True,
-    )
-
-
-@pytest.mark.parametrize("test", pytest_params(WRITE_CONCERN_REJECTED_TESTS))
-def test_setFeatureCompatibilityVersion_writeConcern_rejected(database_client, collection, test):
-    """Test setFeatureCompatibilityVersion rejects non-object writeConcern types."""
     collection = test.prepare(database_client, collection)
     ctx = CommandContext.from_collection(collection)
     cmd = test.build_command(ctx)
